requirements.txt
fastapi==0.68.1
python-multipart==0.0.5
uvicorn[standard]==0.15.0
tensorflow==2.7.0
Pillow==9.0.1
numpy==1.22.1
gunicorn==20.1.0
